requirements.txt
streamlit==1.47.0
yfinance==0.2.65
pandas==2.3.1
numpy==2.2.6
matplotlib==3.10.3
plotly==5.17.0
seaborn==0.13.2
scikit-learn==1.5.0
ta==0.11.0
requests==2.32.4
beautifulsoup4==4.13.4
python-dotenv==1.1.1
tradingview-ta==3.3.0
googlesearch-python==1.2.5
textblob==0.18.0
newsapi-python==0.2.7
xgboost==2.1.3
lightgbm==4.5.0
websocket-client==1.8.0
apscheduler==3.10.4
